feat: enabled support for multiple default drives
