+++ .env
@@ -1,5 +1,5 @@
 VITE_DISABLED_EDITORS=powerhouse/document-drive
-VITE_DEFAULT_DRIVE_URL=https://apps.powerhouse.io/alpha/makerdao/switchboard/d/monetalis
+VITE_DEFAULT_DRIVES_URL=https://apps.powerhouse.io/alpha/makerdao/switchboard/d/monetalis
 VITE_BASE_HREF=/
 ASSET_URL=/
