+++ .github/workflows/build-and-deploy-arbitrum-prod.yaml
@@ -36,7 +36,7 @@ jobs:
             --build-arg VITE_DISABLE_DELETE_CLOUD_DRIVES=true
             --build-arg VITE_DISABLE_DELETE_LOCAL_DRIVES=true
             --build-arg VITE_LOCAL_DRIVES_ENABLED=false
-            --build-arg VITE_DEFAULT_DRIVE_URL=https://apps.powerhouse.io/arbitrum/switchboard/d/arbitrum
+            --build-arg VITE_DEFAULT_DRIVES_URL=https://apps.powerhouse.io/arbitrum/switchboard/d/arbitrum
             --build-arg VITE_RENOWN_CHAIN_ID=42161
             --build-arg VITE_HIDE_DOCUMENT_MODEL_SELECTION_SETTINGS=true
           process_type: web
