+++ Dockerfile
@@ -20,8 +20,8 @@ ENV VITE_SENTRY_DSN=${VITE_SENTRY_DSN}
 ARG VITE_SENTRY_ENV="dev"
 ENV VITE_SENTRY_ENV=${VITE_SENTRY_ENV}
 
-ARG VITE_DEFAULT_DRIVE_URL=""
-ENV VITE_DEFAULT_DRIVE_URL=${VITE_DEFAULT_DRIVE_URL}
+ARG VITE_DEFAULT_DRIVES_URL=""
+ENV VITE_DEFAULT_DRIVES_URL=${VITE_DEFAULT_DRIVES_URL}
 
 ARG VITE_ENABLED_EDITORS="*"
 ENV VITE_ENABLED_EDITORS=${VITE_ENABLED_EDITORS}
